Move client control_protocol into `client.rs`
